Implement Accounts base logic

+++ src/Controllers/ExpenseController.cs
@@ -19,7 +19,16 @@ public class ExpenseController : ApiControllerBase
     [HttpPost]
     public async Task<ActionResult<ExpenseResponse>> Create(CreateExpenseDto createExpenseDto)
     {
-        DateTime now = DateTime.Now;
+        int accountId = createExpenseDto.AccountId;
+        // Todo: Extract into class
+        Account? account = await _dbContext.Accounts.FirstOrDefaultAsync(a => a.Id == accountId);
+        if (account is null)
+        {
+            return NotFound($"Account by id {accountId} not found!");
+        }
+
+        account.Balance -= createExpenseDto.Amount;
+        
         Expense expense = new Expense
         {
             Amount = createExpenseDto.Amount,
@@ -27,9 +36,7 @@ public class ExpenseController : ApiControllerBase
             Note = createExpenseDto.Note,
             CategoryId = createExpenseDto.CategoryId,
             CategoryType = createExpenseDto.CategoryType,
-            MethodId = createExpenseDto.MethodId,
-            CreatedDate = now,
-            UpdatedDate = now,
+            AccountId = accountId,
         };
 
         _dbContext.Expenses.Add(expense);
@@ -41,21 +48,31 @@ public class ExpenseController : ApiControllerBase
     [HttpPut("{id}")]
     public async Task<ActionResult<ExpenseResponse>> Update(int id, UpdateExpenseDto updateExpenseDto)
     {
-        Expense? expense = await _dbContext.Expenses.FindAsync(id);
+        Expense? expense = await _dbContext.Expenses.FirstOrDefaultAsync(e => e.Id == id);
         if (expense is null)
         {
             return NotFound($"Expense with id {id} not found!");
         }
+
+        int accountId = updateExpenseDto.AccountId;
+        // Todo: Extract into class
+        Account? account = await _dbContext.Accounts.FirstOrDefaultAsync(a => a.Id == accountId);
+        if (account is null)
+        {
+            return NotFound($"Account by id {accountId} not found!");
+        }
+
+        decimal amountDifference = updateExpenseDto.Amount - expense.Amount;
+        account.Balance -= amountDifference;
 
         expense.Amount = updateExpenseDto.Amount;
         expense.Date = updateExpenseDto.Date;
         expense.Note = updateExpenseDto.Note;
         expense.CategoryId = updateExpenseDto.CategoryId;
         expense.CategoryType = updateExpenseDto.CategoryType;
-        expense.MethodId = updateExpenseDto.MethodId;
+        expense.AccountId = accountId;
         expense.UpdatedDate = DateTime.Now;
 
-        _dbContext.Expenses.Update(expense);
         await _dbContext.SaveChangesAsync();
 
         return Ok(new ExpenseResponse(expense));
@@ -64,11 +81,21 @@ public class ExpenseController : ApiControllerBase
     [HttpDelete("{id}")]
     public async Task<ActionResult> Delete(int id)
     {
-        Expense? expense = await _dbContext.Expenses.FindAsync(id);
+        Expense? expense = await _dbContext.Expenses.FirstOrDefaultAsync(e => e.Id == id);
         if (expense is null)
         {
             return NotFound($"Expense with id {id} not found!");
         }
+
+        // Todo: Extract into class
+        int accountId = expense.AccountId;
+        Account? account = await _dbContext.Accounts.FirstOrDefaultAsync(a => a.Id == accountId);
+        if (account is null)
+        {
+            return NotFound($"Account by id {accountId} not found!");
+        }
+
+        account.Balance += expense.Amount;
 
         _dbContext.Expenses.Remove(expense);
         await _dbContext.SaveChangesAsync();
@@ -79,7 +106,7 @@ public class ExpenseController : ApiControllerBase
     [HttpGet("{id}")]
     public async Task<ActionResult<ExpenseResponse>> GetById(int id)
     {
-        Expense? expense = await _dbContext.Expenses.FindAsync(id);
+        Expense? expense = await _dbContext.Expenses.FirstOrDefaultAsync(e => e.Id == id);
         if (expense is null)
         {
             return NotFound($"Expense with id {id} not found!");
