+++ src/Controllers/IncomeController.cs
@@ -19,7 +19,16 @@ public class IncomeController : ApiControllerBase
     [HttpPost]
     public async Task<ActionResult<IncomeResponse>> Create(CreateIncomeDto createIncomeDto)
     {
-        DateTime now = DateTime.Now;
+        int accountId = createIncomeDto.AccountId;
+        // Todo: Extract into class
+        Account? account = await _dbContext.Accounts.FirstOrDefaultAsync(a => a.Id == accountId);
+        if (account is null)
+        {
+            return NotFound($"Account by id {accountId} not found!");
+        }
+
+        account.Balance += createIncomeDto.Amount;
+
         Income income = new Income
         {
             Amount = createIncomeDto.Amount,
@@ -27,9 +36,7 @@ public class IncomeController : ApiControllerBase
             Note = createIncomeDto.Note,
             CategoryId = createIncomeDto.CategoryId,
             CategoryType = createIncomeDto.CategoryType,
-            MethodId = createIncomeDto.MethodId,
-            CreatedDate = now,
-            UpdatedDate = now,
+            AccountId = createIncomeDto.AccountId,
         };
 
         _dbContext.Incomes.Add(income);
@@ -41,21 +48,32 @@ public class IncomeController : ApiControllerBase
     [HttpPut("{id}")]
     public async Task<ActionResult<IncomeResponse>> Update(int id, UpdateIncomeDto updateIncomeDto)
     {
-        Income? income = await _dbContext.Incomes.FindAsync(id);
+        Income? income = await _dbContext.Incomes.FirstOrDefaultAsync(i => i.Id == id);
         if (income is null)
         {
             return NotFound($"Income with id {id} not found!");
         }
+
+        // Todo: Extract into class
+        int accountId = updateIncomeDto.AccountId;
+        Account? account = await _dbContext.Accounts.FirstOrDefaultAsync(a => a.Id == accountId);
+        if (account is null)
+        {
+            return NotFound($"Account by id {accountId} not found!");
+        }
+
+        account.Balance -= updateIncomeDto.Amount;
+        decimal amountDifference = updateIncomeDto.Amount - income.Amount;
+        account.Balance += amountDifference;
 
         income.Amount = updateIncomeDto.Amount;
         income.Date = updateIncomeDto.Date;
         income.Note = updateIncomeDto.Note;
         income.CategoryId = updateIncomeDto.CategoryId;
         income.CategoryType = updateIncomeDto.CategoryType;
-        income.MethodId = updateIncomeDto.MethodId;
+        income.AccountId = updateIncomeDto.AccountId;
         income.UpdatedDate = DateTime.Now;
 
-        _dbContext.Incomes.Update(income);
         await _dbContext.SaveChangesAsync();
 
         return Ok(new IncomeResponse(income));
@@ -65,11 +83,20 @@ public class IncomeController : ApiControllerBase
     public async Task<ActionResult> Delete(int id)
     {
         Income? income = await _dbContext.Incomes.FirstOrDefaultAsync(i => i.Id == id);
-
         if (income is null)
         {
             return NotFound();
         }
+
+        // Todo: Extract into class
+        int accountId = income.AccountId;
+        Account? account = await _dbContext.Accounts.FirstOrDefaultAsync(a => a.Id == accountId);
+        if (account is null)
+        {
+            return NotFound($"Account by id {accountId} not found!");
+        }
+
+        account.Balance -= income.Amount;
 
         _dbContext.Incomes.Remove(income);
         await _dbContext.SaveChangesAsync();
@@ -80,7 +107,7 @@ public class IncomeController : ApiControllerBase
     [HttpGet("{id}")]
     public async Task<ActionResult<IncomeResponse>> GetById(int id)
     {
-        Income? income = await _dbContext.Incomes.FindAsync(id);
+        Income? income = await _dbContext.Incomes.FirstOrDefaultAsync(i => i.Id == id);
         if (income is null)
         {
             return NotFound($"Income with id {id} not found!");
